pyproject.toml
[tool.poetry]
name = "omega-miya"
version = "1.0.1"
description = "基于 nonebot2 的多平台机器人"
authors = ["Ailitonia <41713304+Ailitonia@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = "^3.12"
nonebot2 = { version = "2.4.2", extras = ["fastapi", "aiohttp", "httpx", "websockets"] }
pydantic = ">=2.11.0,<3.0.0"
nonebot-adapter-console = ">=0.6.0,<1.0.0"
nonebot-adapter-onebot = ">=2.4.6,<2.5.0"
nonebot-adapter-qq = ">=1.6.4,<1.7.0"
nonebot-adapter-telegram = ">=0.1.0b20,<0.2.0"
nonebot-adapter-feishu = ">=2.6.0,<2.7.0"
nonebot-adapter-discord = ">=0.1.8,<1.0.0"
nonebot-plugin-apscheduler = ">=0.5.0,<1.0.0"
sqlalchemy = ">=2.0.40,<2.1.0"
asyncmy = { version = ">=0.2.10,<1.0.0", optional = true }
aiomysql = { version = ">=0.2.0,<1.0.0", optional = true }
asyncpg = { version = ">=0.30.0,<1.0.0", optional = true }
aiosqlite = { version = ">=0.21.0,<1.0.0", optional = true }
apscheduler = ">=3.11.0,<4.0.0"
aiofiles = ">=24.1.0,<25.0.0"
ujson = ">=5.10.0,<6.0.0"
lxml = ">=5.4.0,<6.0.0"
xmltodict = ">=0.14.2,<1.0.0"
msgpack = ">=1.0.8,<2.0.0"
numpy = ">=2.2.0,<3.0.0"
faiss-cpu = ">=1.11.0,<2.0.0"
openpyxl = ">=3.1.5,<4.0.0"
pandas = ">=2.2.0,<3.0.0"
matplotlib = ">=3.10.0,<4.0.0"
pillow = ">=11.2.0,<12.0.0"
imageio = ">=2.37.0,<3.0.0"
fonttools = ">=4.58.0,<5.0.0"
psutil = ">=7.0.0,<8.0.0"
pycryptodome = ">=3.23.0,<4.0.0"
py7zr = ">=1.0.0,<2.0.0"
pytz = "^2025.2"
zhconv = ">=1.4.3,<2.0.0"
rapidfuzz = ">=3.13.0,<4.0.0"
emoji = ">=2.14.0,<3.0.0"
qrcode = { extras = ["pil"], version = ">=8.0.0,<9.0.0" }
onedice = "1.0.7"
jieba = ">=0.42.1,<1.0.0"
wordcloud = ">=1.9.4,<2.0.0"


[tool.poetry.extras]
mysql = ["asyncmy", "aiomysql"]
pgsql = ["asyncpg"]
sqlite = ["aiosqlite"]


[tool.poetry.group.dev.dependencies]
mypy = ">=1.16.0,<2.0.0"
ruff = ">=0.11.11,<1.0.0"
bump-pydantic = "^0.8.0"
types-lxml = "^2025.3.30"
lxml-stubs = "^0.5.1"
alembic = "^1.16.0"
memory-profiler = ">=0.61.0,<1.0.0"
pywebio = ">=1.8.4,<2.0.0"
pywebio-battery = ">=0.7.0,<1.0.0"
Jinja2 = ">=3.1.5,<4.0.0"


[[tool.poetry.source]]
name = "清华"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"


[[tool.poetry.source]]
name = "中国科大"
url = "https://mirrors.ustc.edu.cn/pypi/simple/"
priority = "explicit"


[[tool.poetry.source]]
name = "阿里云"
url = "https://mirrors.aliyun.com/pypi/simple/"
priority = "explicit"


[[tool.poetry.source]]
name = "腾讯云"
url = "https://mirrors.cloud.tencent.com/pypi/simple/"
priority = "explicit"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.flake8]
max-line-length = 120
ignore = [
    "W503",
]


[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".github",
    ".hg",
    ".idea",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]
line-length = 120
indent-width = 4
target-version = "py312"  # Assume Python 3.12


[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = [
    "F", # Pyflakes
    "E", # pycodestyle Error (E)
    "W", # pycodestyle Warning (W)
    "I001", # unsorted-imports
    "UP", # pyupgrade (UP)
    "C4", # flake8-comprehensions (C4)
    "T10", # flake8-debugger (T10)
    "T20", # flake8-print (T20)
    "PYI", # flake8-pyi (PYI)
    "PT", # flake8-pytest-style (PT)
    "Q", # flake8-quotes (Q)
    "NPY", # NumPy-specific rules (NPY)
    "FAST", # FastAPI (FAST)
]
ignore = [
    "E402", # module-import-not-at-top-of-file
    "C901", # complex-structure
    "PT023", # pytest-incorrect-mark-parentheses-style
]
isort.case-sensitive = true
flake8-quotes.inline-quotes = "single"


[tool.ruff.format]
quote-style = "single"  # Use single quotes for strings.
indent-style = "space"  # Indent with spaces, rather than tabs.
